backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-telegram-bot[job-queue]==21.1.1
langchain==0.2.1
langchain-openai==0.1.7
pydantic==2.7.1
docker==7.0.0
websockets==12.0
python-dotenv==1.0.1
networkx==3.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Blockchain (EAS)
web3==6.15.1
eth-account==0.11.0

# Additional
httpx==0.27.0
redis==5.0.3
